Update of the counter class

+++ FlowMate_Team5/src/flowmate_team5/models/Counter.java
@@ -3,22 +3,22 @@ package flowmate_team5.models;
 import java.io.Serializable;
 
 /**
- * Represents a counter entity with a name and a value.
- * Implements Serializable to allow saving/loading state if necessary.
+ * Represents a counter entity with a name and an integer value.
+ * Updated to use int as per requirements.
  */
 public class Counter implements Serializable {
 
     private static final long serialVersionUID = 1L;
     private String name;
-    private double value;
+    private int value;
 
-    public Counter(String name, double initialValue) {
+    public Counter(String name, int initialValue) {
         this.name = name;
         this.value = initialValue;
     }
 
     public Counter(String name) {
-        this(name, 0.0);
+        this(name, 0);
     }
 
     public void increment() {
@@ -29,11 +29,11 @@ public class Counter implements Serializable {
         this.value--;
     }
 
-    public double getValue() {
+    public int getValue() {
         return value;
     }
 
-    public void setValue(double value) {
+    public void setValue(int value) {
         this.value = value;
     }
